test: run live compaction reload journey
